Move x509_crl_info code to module_utils.
